scripts for managing large images

+++ citify.sh
@@ -4,6 +4,7 @@
 # metadata is prepended to any existing file content
 
 app=${1}
+INIT=.init.md
 
 find ./ -type f -iname \*.md -print0 | sort -z | while read -d $'\0' file; do
 
@@ -11,28 +12,37 @@ find ./ -type f -iname \*.md -print0 | sort -z | while read -d $'\0' file; do
     # check first line for front matter start string: "---"
     [ $(sed -n '1{/^\-\-\-/p};q' "${file}") ] && { echo "  - ${file} already has metadata defined, skipping."; continue; }
 
-    filename="${file##*/}"
-    basename="${filename%.*}"
-   
-    citetext="$basename"
-    date=`date -d "${citetext:0:10}" +'%d %b %Y' 2> /dev/null` 
-    datevalidity=$? # get date command exit status
-    # check that we have a valid ISO date
-    if [ "${datevalidity}" -eq 0 ]; then
-        citetext="${date}, ${citetext:10}"
-    fi
-   
-    # replace dashes with spaces, title case, and append any command line argument
-    citetext="${citetext//-/ }"
-    # citetext=`echo "${citetext}" | sed 's/[^ ]\+/\L\u&/g'`
-    if [ -n "${app}" ]; then
-        citetext="${citetext} ${app}"  
-    fi
-    
-    # prepend formatted front matter string to file
-    mdcontent="---\ncitation: \"${citetext}.\"\n---\n\n"
-    mdcontent+=`cat "${file}"`
-    echo -e $mdcontent > $file
+    if [[ -f "$INIT" ]]; then
 
+        # overwrite without confirmation, which should be safe now...
+        \cp ${INIT} ${file}
+
+    else 
+
+        filename="${file##*/}"
+        basename="${filename%.*}"
+       
+        citetext="$basename"
+        date=`date -d "${citetext:0:10}" +'%d %b %Y' 2> /dev/null` 
+        datevalidity=$? # get date command exit status
+        # check that we have a valid ISO date
+        if [ "${datevalidity}" -eq 0 ]; then
+            citetext="${date}, ${citetext:10}"
+        fi
+       
+        # replace dashes with spaces, title case, and append any command line argument
+        citetext="${citetext//-/ }"
+        # citetext=`echo "${citetext}" | sed 's/[^ ]\+/\L\u&/g'`
+        if [ -n "${app}" ]; then
+            citetext="${citetext} ${app}"  
+        fi
+        
+        # prepend formatted front matter string to file
+        mdcontent="---\ncitation: \"${citetext}.\"\n---\n\n"
+        mdcontent+=`cat "${file}"`
+        echo -e $mdcontent > $file
+
+    fi
 done
 
+mv .init.md .init.bac
